requirements.txt
Django
django-extensions
django-webpack-loader
django-sesame
python-telegram-bot
django-environ